+++ src/constants.ts
@@ -1,5 +1,6 @@
 export namespace commmands {
     export const generateParameters = "arm.generateParameterFile";
+    export const consolidateFiles = "arm.consolidateFiles";
 }
 
 export namespace config {