+++ src/extension.ts
@@ -3,15 +3,13 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 import * as constants from './constants';
-import { ExtensionState } from "./common/extension-state";
 import { generateParameterFile } from './commands/generate-parameters-file';
+import { consolidateFiles } from './commands/consolidate-parameters-file';
 
 export function activate(context: vscode.ExtensionContext) {
-    ExtensionState.configure(context);
-
     context.subscriptions.push(
-        //vscode.commands.registerTextEditorCommand(constants.commmands.generateParameters, generateParameterFile),
-        vscode.commands.registerCommand(constants.commmands.generateParameters, generateParameterFile)
+        vscode.commands.registerCommand(constants.commmands.generateParameters, generateParameterFile),
+        vscode.commands.registerCommand(constants.commmands.consolidateFiles, consolidateFiles)
     );
 }